Change: Move SingleMenu into separate class, hide hidden folders on *Nix
Change: Hide hidden folders on Linux / Mac
