+++ source/OpenBVE/Game/Menu/Menu.cs
@@ -2,9 +2,7 @@ using OpenBveApi.Colors;
 using OpenBveApi.Graphics;
 using OpenBveApi.Interface;
 using System;
-using System.ComponentModel;
 using System.Drawing;
-using DavyKager;
 using System.IO;
 using System.Text;
 using System.Windows.Forms;
@@ -17,7 +15,6 @@ using OpenBveApi.Input;
 using OpenBveApi.Textures;
 using OpenTK;
 using TrainManager;
-using Control = OpenBveApi.Interface.Control;
 using Path = OpenBveApi.Path;
 using Vector2 = OpenBveApi.Math.Vector2;
 
@@ -59,452 +56,7 @@ namespace OpenBve
 
 		private double lastTimeElapsed;
 		
-		/********************
-			SINGLE-MENU ENTRY CLASS
-		*********************
-		Describes a single menu of the menu stack.
-		The class is private to Menu, but all its fields are public to allow 'quick-and-dirty'
-		access from Menu itself. */
-		private class SingleMenu
-		{
-			/********************
-				MENU FIELDS
-			*********************/
-			public readonly TextAlignment Align;
-			public readonly MenuEntry[] Items = { };
-			public readonly double ItemWidth = 0;
-			public readonly double Width = 0;
-			public readonly double Height = 0; 
-			public readonly double MaxWidth;
-
-			private int lastSelection = int.MaxValue;
-			private int currentSelection;
-			
-			public int Selection
-			{
-				get
-				{
-					return currentSelection;
-				}
-				set
-				{
-					lastSelection = currentSelection;
-					currentSelection = value;
-					if (currentSelection != lastSelection && Interface.CurrentOptions.ScreenReaderAvailable)
-					{
-						if (!Tolk.Output(Items[currentSelection].Text))
-						{
-							// failed to output to screen reader, so don't keep trying
-							Interface.CurrentOptions.ScreenReaderAvailable = false;
-						}
-					}
-				}
-			}
-			public int TopItem;         // the top displayed menu item
-			internal readonly MenuType Type;
-			
-
-			/********************
-				MENU C'TOR
-			*********************/
-			public SingleMenu(MenuType menuType, int data = 0, double maxWidth = 0)
-			{
-				MaxWidth = maxWidth;
-				Type = menuType;
-				int i, menuItem;
-				int jump = 0;
-				Vector2 size;
-
-				Align = TextAlignment.TopMiddle;
-				Height = Width = 0;
-				Selection = 0;                      // defaults to first menu item
-				switch (menuType)
-				{
-					case MenuType.GameStart:          // top level menu
-						if (routeWorkerThread == null)
-						{
-							//Create the worker thread for route details processing on first launch of main menu
-							routeWorkerThread = new BackgroundWorker();
-							routeWorkerThread.DoWork += routeWorkerThread_doWork;
-							routeWorkerThread.RunWorkerCompleted += routeWorkerThread_completed;
-							//Load texture
-							Program.CurrentHost.RegisterTexture(Path.CombineFile(Program.FileSystem.DataFolder, "Menu\\loading.png"), new TextureParameters(null, null), out routePictureBox.Texture);
-						}
-						Items = new MenuEntry[3];
-						Items[0] = new MenuCommand("Open Route File", MenuTag.RouteList, 0);
-						
-						if (!Interface.CurrentOptions.KioskMode)
-						{
-							//Don't allow quitting or customisation of the controls in kiosk mode
-							Items[1] = new MenuCommand(Translations.GetInterfaceString("menu_customize_controls"), MenuTag.MenuControls, 0);
-							Items[2] = new MenuCommand(Translations.GetInterfaceString("menu_quit"), MenuTag.MenuQuit, 0);
-						}
-						else
-						{
-							Array.Resize(ref Items, Items.Length - 2);
-						}
-						SearchDirectory = Program.FileSystem.InitialRouteFolder;
-						Align = TextAlignment.TopLeft;
-						break;
-					case MenuType.RouteList:
-						string[] potentialFiles = { };
-						string[] directoryList = { };
-						bool drives = false;
-						if (SearchDirectory != string.Empty)
-						{
-							try
-							{
-								potentialFiles = Directory.GetFiles(SearchDirectory);
-								directoryList = Directory.GetDirectories(SearchDirectory);
-							}
-							catch
-							{
-								// Ignored
-							}
-						}
-						else
-						{
-							DriveInfo[] systemDrives = DriveInfo.GetDrives();
-							directoryList = new string[systemDrives.Length];
-							for (int k = 0; k < systemDrives.Length; k++)
-							{
-								directoryList[k] = systemDrives[k].Name;
-							}
-							drives = true;
-						}
-						
-						Items = new MenuEntry[potentialFiles.Length + directoryList.Length + 2];
-						Items[0] = new MenuCaption(SearchDirectory);
-						Items[1] = new MenuCommand("...", MenuTag.ParentDirectory, 0);
-						int totalEntries = 2;
-						for (int j = 0; j < directoryList.Length; j++)
-						{
-							Items[totalEntries] = new MenuCommand(new DirectoryInfo(directoryList[j]).Name, MenuTag.Directory, 0);
-							if (drives)
-							{
-								Program.CurrentHost.RegisterTexture(Path.CombineFile(Program.FileSystem.DataFolder, "Menu\\icon_disk.png"), new TextureParameters(null, null), out Items[totalEntries].Icon);
-							}
-							else
-							{
-								Program.CurrentHost.RegisterTexture(Path.CombineFile(Program.FileSystem.DataFolder, "Menu\\icon_folder.png"), new TextureParameters(null, null), out Items[totalEntries].Icon);	
-							}
-							
-							totalEntries++;
-						}
-
-						for (int j = 0; j < potentialFiles.Length; j++)
-						{
-							string fileName = System.IO.Path.GetFileName(potentialFiles[j]);
-							if (fileName.ToLowerInvariant().EndsWith(".csv") || fileName.ToLowerInvariant().EndsWith(".rw"))
-							{
-								Items[totalEntries] = new MenuCommand(fileName, MenuTag.RouteFile, 0);
-								Program.CurrentHost.RegisterTexture(Path.CombineFile(Program.FileSystem.DataFolder, "Menu\\icon_route.png"), new TextureParameters(null, null), out Items[totalEntries].Icon);
-								totalEntries++;
-							}
-						}
-						Array.Resize(ref Items, totalEntries);
-						Align = TextAlignment.TopLeft;
-						break;
-					case MenuType.TrainList:
-						potentialFiles = new string[] { };
-						directoryList = new string[] { };
-						drives = false;
-						if (SearchDirectory != string.Empty)
-						{
-							try
-							{
-								potentialFiles = Directory.GetFiles(SearchDirectory);
-								directoryList = Directory.GetDirectories(SearchDirectory);
-							}
-							catch
-							{
-								// Ignored
-							}
-						}
-						else
-						{
-							DriveInfo[] systemDrives = DriveInfo.GetDrives();
-							directoryList = new string[systemDrives.Length];
-							for (int k = 0; k < systemDrives.Length; k++)
-							{
-								directoryList[k] = systemDrives[k].Name;
-							}
-							drives = true;
-						}
-						
-						Items = new MenuEntry[potentialFiles.Length + directoryList.Length + 2];
-						Items[0] = new MenuCaption(SearchDirectory);
-						Items[1] = new MenuCommand("...", MenuTag.ParentDirectory, 0);
-						totalEntries = 2;
-						for (int j = 0; j < directoryList.Length; j++)
-						{
-							bool isTrain = false;
-							for (int k = 0; k < Program.CurrentHost.Plugins.Length; k++)
-							{
-								if (Program.CurrentHost.Plugins[k].Train != null && Program.CurrentHost.Plugins[k].Train.CanLoadTrain(directoryList[j]))
-								{
-									isTrain = true;
-									break;
-								}
-							}
-
-							if (!isTrain)
-							{
-								Items[totalEntries] = new MenuCommand(new DirectoryInfo(directoryList[j]).Name, MenuTag.Directory, 0);
-								if (drives)
-								{
-									Program.CurrentHost.RegisterTexture(Path.CombineFile(Program.FileSystem.DataFolder, "Menu\\icon_disk.png"), new TextureParameters(null, null), out Items[totalEntries].Icon);
-								}
-								else
-								{
-									Program.CurrentHost.RegisterTexture(Path.CombineFile(Program.FileSystem.DataFolder, "Menu\\icon_folder.png"), new TextureParameters(null, null), out Items[totalEntries].Icon);	
-								}
-							}
-							else
-							{
-								Items[totalEntries] = new MenuCommand(new DirectoryInfo(directoryList[j]).Name, MenuTag.TrainDirectory, 0);
-								Program.CurrentHost.RegisterTexture(Path.CombineFile(Program.FileSystem.DataFolder, "Menu\\icon_train.png"), new TextureParameters(null, null), out Items[totalEntries].Icon);	
-							}
-							totalEntries++;
-						}
-						Array.Resize(ref Items, totalEntries);
-						Align = TextAlignment.TopLeft;
-						break;
-					case MenuType.Top:          // top level menu
-						if (Interface.CurrentOptions.ScreenReaderAvailable)
-						{
-							if (!Tolk.Output(Translations.GetInterfaceString("menu_title")))
-							{
-								// failed to output to screen reader, so don't keep trying
-								Interface.CurrentOptions.ScreenReaderAvailable = false;
-							}
-						}
-						for (i = 0; i < Program.CurrentRoute.Stations.Length; i++)
-							if (Program.CurrentRoute.Stations[i].PlayerStops() & Program.CurrentRoute.Stations[i].Stops.Length > 0)
-							{
-								jump = 1;
-								break;
-							}
-						Items = new MenuEntry[4 + jump];
-						Items[0] = new MenuCommand(Translations.GetInterfaceString("menu_resume"), MenuTag.BackToSim, 0);
-						if (jump > 0)
-							Items[1] = new MenuCommand(Translations.GetInterfaceString("menu_jump"), MenuTag.MenuJumpToStation, 0);
-						if (!Interface.CurrentOptions.KioskMode)
-						{
-							//Don't allow quitting or customisation of the controls in kiosk mode
-							Items[1 + jump] = new MenuCommand(Translations.GetInterfaceString("menu_exit"), MenuTag.MenuExitToMainMenu, 0);
-							Items[2 + jump] = new MenuCommand(Translations.GetInterfaceString("menu_customize_controls"), MenuTag.MenuControls, 0);
-							Items[3 + jump] = new MenuCommand(Translations.GetInterfaceString("menu_quit"), MenuTag.MenuQuit, 0);
-						}
-						else
-						{
-							Array.Resize(ref Items, Items.Length -3);
-						}
-						break;
-					case MenuType.JumpToStation:    // list of stations to jump to
-													// count the number of available stations
-						menuItem = 0;
-						for (i = 0; i < Program.CurrentRoute.Stations.Length; i++)
-							if (Program.CurrentRoute.Stations[i].PlayerStops() & Program.CurrentRoute.Stations[i].Stops.Length > 0)
-								menuItem++;
-						// list available stations, selecting the next station as predefined choice
-						jump = 0;                           // no jump found yet
-						Items = new MenuEntry[menuItem + 1];
-						Items[0] = new MenuCommand(Translations.GetInterfaceString("menu_back"), MenuTag.MenuBack, 0);
-						menuItem = 1;
-						for (i = 0; i < Program.CurrentRoute.Stations.Length; i++)
-							if (Program.CurrentRoute.Stations[i].PlayerStops() & Program.CurrentRoute.Stations[i].Stops.Length > 0)
-							{
-								Items[menuItem] = new MenuCommand(Program.CurrentRoute.Stations[i].Name, MenuTag.JumpToStation, i);
-								// if no preferred jump-to-station found yet and this station is
-								// after the last station the user stopped at, select this item
-								if (jump == 0 && i > TrainManagerBase.PlayerTrain.LastStation)
-								{
-									jump = i;
-									Selection = menuItem;
-								}
-								menuItem++;
-							}
-						Align = TextAlignment.TopLeft;
-						break;
-
-					case MenuType.ExitToMainMenu:
-						Items = new MenuEntry[3];
-						Items[0] = new MenuCaption(Translations.GetInterfaceString("menu_exit_question"));
-						Items[1] = new MenuCommand(Translations.GetInterfaceString("menu_exit_no"), MenuTag.MenuBack, 0);
-						Items[2] = new MenuCommand(Translations.GetInterfaceString("menu_exit_yes"), MenuTag.ExitToMainMenu, 0);
-						Selection = 1;
-						break;
-
-					case MenuType.Quit:         // ask for quit confirmation
-						Items = new MenuEntry[3];
-						Items[0] = new MenuCaption(Translations.GetInterfaceString("menu_quit_question"));
-						Items[1] = new MenuCommand(Translations.GetInterfaceString("menu_quit_no"), MenuTag.MenuBack, 0);
-						Items[2] = new MenuCommand(Translations.GetInterfaceString("menu_quit_yes"), MenuTag.Quit, 0);
-						Selection = 1;
-						break;
-
-					case MenuType.Controls:
-						//Refresh the joystick list
-						Program.Joysticks.RefreshJoysticks();
-						Items = new MenuEntry[Interface.CurrentControls.Length + 1];
-						Items[0] = new MenuCommand(Translations.GetInterfaceString("menu_back"), MenuTag.MenuBack, 0);
-						for (i = 0; i < Interface.CurrentControls.Length; i++)
-							Items[i + 1] = new MenuCommand(Interface.CurrentControls[i].Command.ToString(), MenuTag.Control, i);
-						Align = TextAlignment.TopLeft;
-						break;
-
-					case MenuType.Control:
-						//Refresh the joystick list
-						Program.Joysticks.RefreshJoysticks();
-						Selection = SelectionNone;
-						Items = new MenuEntry[4];
-						// get code name and description
-						Control loadedControl = Interface.CurrentControls[data];
-						for (int h = 0; h < Translations.CommandInfos.Length; h++)
-						{
-							if (Translations.CommandInfos[h].Command == loadedControl.Command)
-							{
-								Items[0] = new MenuCommand(loadedControl.Command.ToString() + " - " +
-										Translations.CommandInfos[h].Description, MenuTag.None, 0);
-								break;
-							}
-						}
-						// get assignment
-						String str = "";
-						switch (loadedControl.Method)
-						{
-							case ControlMethod.Keyboard:
-								string keyName = loadedControl.Key.ToString();
-								for (int k = 0; k < Translations.TranslatedKeys.Length; k++)
-								{
-									if (Translations.TranslatedKeys[k].Key == loadedControl.Key)
-									{
-										keyName = Translations.TranslatedKeys[k].Description;
-										break;
-									}
-								}
-								if (loadedControl.Modifier != KeyboardModifier.None)
-								{
-									str = Translations.GetInterfaceString("menu_keyboard") + " [" + loadedControl.Modifier + "-" + keyName + "]";
-								}
-								else
-								{
-									str = Translations.GetInterfaceString("menu_keyboard") + " [" + keyName + "]";
-								}
-								break;
-							case ControlMethod.Joystick:
-								str = Translations.GetInterfaceString("menu_joystick") + " " + loadedControl.Device + " [" + loadedControl.Component + " " + loadedControl.Element + "]";
-								switch (loadedControl.Component)
-								{
-									case JoystickComponent.FullAxis:
-									case JoystickComponent.Axis:
-										str += " " + (loadedControl.Direction == 1 ? Translations.GetInterfaceString("menu_joystickdirection_positive") : Translations.GetInterfaceString("menu_joystickdirection_negative"));
-										break;
-									//						case Interface.JoystickComponent.Button:	// NOTHING TO DO FOR THIS CASE!
-									//							str = str;
-									//							break;
-									case JoystickComponent.Hat:
-										str += " " + (OpenTK.Input.HatPosition)loadedControl.Direction;
-										break;
-									case JoystickComponent.Invalid:
-										str = Translations.GetInterfaceString("menu_joystick_notavailable");
-										break;
-								}
-								break;
-							case ControlMethod.RailDriver:
-								str = "RailDriver [" + loadedControl.Component + " " + loadedControl.Element + "]";
-								switch (loadedControl.Component)
-								{
-									case JoystickComponent.FullAxis:
-									case JoystickComponent.Axis:
-										str += " " + (loadedControl.Direction == 1 ? Translations.GetInterfaceString("menu_joystickdirection_positive") : Translations.GetInterfaceString("menu_joystickdirection_negative"));
-										break;
-									case JoystickComponent.Invalid:
-										str = Translations.GetInterfaceString("menu_joystick_notavailable");
-										break;
-								}
-								break;
-							case ControlMethod.Invalid:
-								str = Translations.GetInterfaceString("menu_joystick_notavailable");
-								break;
-						}
-						Items[1] = new MenuCommand(Translations.GetInterfaceString("menu_assignment_current") + " " + str, MenuTag.None, 0);
-						Items[2] = new MenuCommand(" ", MenuTag.None, 0);
-						Items[3] = new MenuCommand(Translations.GetInterfaceString("menu_assign"), MenuTag.None, 0);
-						break;
-					case MenuType.TrainDefault:
-						Interface.CurrentOptions.TrainFolder = Loading.GetDefaultTrainFolder(RouteFile);
-						bool canLoad = false;
-						for (int j = 0; j < Program.CurrentHost.Plugins.Length; j++)
-						{
-							if (Program.CurrentHost.Plugins[j].Train != null && Program.CurrentHost.Plugins[j].Train.CanLoadTrain(Interface.CurrentOptions.TrainFolder))
-							{
-								canLoad = true;
-								break;
-							}
-						}
-
-						if (canLoad)
-						{
-							Items = new MenuEntry[3];
-							Items[0] = new MenuCaption(Translations.GetInterfaceString("start_train_default"));
-							Items[1] = new MenuCommand(Translations.GetInterfaceString("start_train_default_yes"), MenuTag.Yes, 0);
-							Items[2] = new MenuCommand(Translations.GetInterfaceString("start_train_default_no"), MenuTag.No, 0);
-							Selection = 1;
-						}
-						else
-						{
-							SearchDirectory = Program.FileSystem.InitialTrainFolder;
-							//Default train not found or not valid
-							Instance.PushMenu(MenuType.TrainList);
-						}
-						break;
-				}
-				// compute menu extent
-				for (i = 0; i < Items.Length; i++)
-				{
-					if (Items[i] == null)
-					{
-						continue;
-					}
-					size = Game.Menu.MenuFont.MeasureString(Items[i].Text);
-					if (Items[i].Icon != null)
-					{
-						size.X += size.Y * 1.25;
-					}
-					if (size.X > Width)
-					{
-						Width = size.X;
-					}
-					
-					if (MaxWidth != 0 && size.X > MaxWidth)
-					{
-						for (int j = Items[i].Text.Length - 1; j > 0; j--)
-						{
-							string trimmedText = Items[i].Text.Substring(0, j);
-							size = Game.Menu.MenuFont.MeasureString(trimmedText);
-							double mwi = MaxWidth;
-							if (Items[i].Icon != null)
-							{
-								mwi -= size.Y * 1.25;
-							}
-							if (size.X < mwi)
-							{
-								Items[i].DisplayLength = trimmedText.Length;
-								break;
-							}
-						}
-						Width = MaxWidth;
-					}
-					if (!(Items[i] is MenuCaption && menuType!= MenuType.RouteList && menuType != MenuType.GameStart) && size.X > ItemWidth)
-						ItemWidth = size.X;
-				}
-				Height = Items.Length * Game.Menu.LineHeight;
-				TopItem = 0;
-			}
-
-		}                   // end of private class SingleMenu
+		                   // end of private class SingleMenu
 
 		/********************
 			MENU SYSTEM FIELDS
