+++ source/OpenBVE/OpenBve.csproj
@@ -141,6 +141,7 @@
     <Compile Include="Game\Menu\Menu.MenuEntry.cs" />
     <Compile Include="Game\Menu\Menu.Route.cs" />
     <Compile Include="Game\Menu\Menu.RouteState.cs" />
+    <Compile Include="Game\Menu\Menu.SingleMenu.cs" />
     <Compile Include="Game\Menu\MenuTag.cs" />
     <Compile Include="Game\Menu\MenuType.cs" />
     <Compile Include="Game\MessageManager.cs" />
